counter_t op int32t gezet voor versnelling (in ieder geval op macos)

+++ PrimeC/solution_5/sieve_helpers.h
@@ -139,7 +139,7 @@
 #endif
 
 #define bitshift_t uint64_t // type used to shift bits
-#define counter_t  uint64_t // type used to count loops, etc
+#define counter_t  int32_t // type used to count loops, etc
 
 // masks and mask helpers
 #define SHIFT_BYTE          3
